Handle and forward errors instead of unwrapping all over the place
