+++ datastore_entity/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["Andreas Pålsson <a.palsson@gmail.com>", "Erik Zivkovic <erik@zivkov
 edition = "2018"
 
 [dependencies]
+thiserror = "1.0.22"
 google-datastore1 = { git = "https://github.com/bes/google-datastore1-generated" }
 google_api_auth = { git = "https://github.com/google-apis-rs/generator" }
 gcp_auth = "0.3.2"
